move delete command to archiver.delete
